[Test][Autoscaler] Rename e2e_autoscaler to e2eautoscaler to make the linter happy
