+++ .buildkite/test-sample-yamls.yml
@@ -64,7 +64,9 @@
   instance_size: large
   image: golang:1.20
   commands:
-    - ./.buildkite/setup-env.sh
+    - source .buildkite/setup-env.sh
+    - kind create cluster --wait 900s --config ./tests/framework/config/kind-config-buildkite.yml
+    - kubectl config set clusters.kind-kind.server https://docker:6443
     # Build nightly KubeRay operator image
     - pushd ray-operator
     - IMG=kuberay/operator:nightly make docker-image
@@ -72,6 +74,6 @@
     - IMG=kuberay/operator:nightly make deploy
     - kubectl wait --timeout=90s --for=condition=Available=true deployment -n ray-system kuberay-operator
     # Run e2e tests
-    - KUBERAY_TEST_TIMEOUT_SHORT=1m KUBERAY_TEST_TIMEOUT_MEDIUM=5m KUBERAY_TEST_TIMEOUT_LONG=10m go test -timeout 30m -v ./test/e2e_autoscaler
+    - KUBERAY_TEST_TIMEOUT_SHORT=1m KUBERAY_TEST_TIMEOUT_MEDIUM=5m KUBERAY_TEST_TIMEOUT_LONG=10m go test -timeout 30m -v ./test/e2eautoscaler
     # Printing KubeRay operator logs
     - kubectl logs -n ray-system --tail -1 -l app.kubernetes.io/name=kuberay
